requirements.txt
opencv-python>=4.5.5
numpy
pytesseract
imutils
fpdf2
